+++ lights/patterns.h
@@ -3,22 +3,22 @@
 
 #include "lib.h"
 
-CRGB pattern0(int frame, int led);
-CRGB pattern1(int frame, int led);
-CRGB pattern2(int frame, int led);
-CRGB pattern3(int frame, int led);
-CRGB pattern4(int frame, int led);
-CRGB pattern5(int frame, int led);
-CRGB pattern6(int frame, int led);
-CRGB pattern7(int frame, int led);
-CRGB pattern8(int frame, int led);
-CRGB pattern9(int frame, int led);
-CRGB pattern10(int frame, int led);
-CRGB pattern11(int frame, int led);
-CRGB pattern12(int frame, int led);
-CRGB pattern13(int frame, int led);
-CRGB pattern14(int frame, int led);
-CRGB pattern15(int frame, int led);
+CRGB pattern0(long frame, int led);
+CRGB pattern1(long frame, int led);
+CRGB pattern2(long frame, int led);
+CRGB pattern3(long frame, int led);
+CRGB pattern4(long frame, int led);
+CRGB pattern5(long frame, int led);
+CRGB pattern6(long frame, int led);
+CRGB pattern7(long frame, int led);
+CRGB pattern8(long frame, int led);
+CRGB pattern9(long frame, int led);
+CRGB pattern10(long frame, int led);
+CRGB pattern11(long frame, int led);
+CRGB pattern12(long frame, int led);
+CRGB pattern13(long frame, int led);
+CRGB pattern14(long frame, int led);
+CRGB pattern15(long frame, int led);
 
 #endif
